src/test/java/com/danilkompaniets/app/student/StudentMapperTest.java
package com.danilkompaniets.app.student;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class StudentMapperTest {
    private StudentMapper mapper;

    @BeforeEach
    void setUp() {
        mapper = new StudentMapper();
    }

    @Test
    public void testMethod1() {
        StudentDto dto = new StudentDto("John", "Doe", "johndoe@gmail.com", 1);
        Student student = mapper.toStudent(dto);

        assertEquals(dto.firstName(), student.getFirstName());
        assertEquals(dto.lastName(), student.getLastName());
        assertEquals(dto.email(), student.getEmail());
        assertNotNull(student.getSchool());
        assertEquals(dto.schoolId(), student.getSchool().getId());
    }

    @Test
    public void shouldMapStudentToStudentResponseDto() {
        Student student = new Student(12, "jsmith@gmail.com", "Jane", "Smith");

        StudentResponseDto response = mapper.toStudentResponseDto(student);
        assertEquals(response.firstName(), student.getFirstName());
        assertEquals(response.lastName(), student.getLastName());
        assertEquals(response.email(), student.getEmail());
    }

    @Test
    public void should_throw_exception_when_mapStudentToStudentResponseDto_null() {
        var exp = assertThrows(NullPointerException.class, () -> mapper.toStudent(null));
        assertEquals("The student dto cannot be null", exp.getMessage());
    }
}